fixed session_restart

+++ core_v2.py
@@ -476,19 +476,26 @@ async def update_leagues():
         query = await supabase.table('discord_osu').select('discord_username, discord_id, league, future_league').execute()
         datas = query.data
         for data in datas:
-            print(data)
             league = data['league']
             future_league = data['future_league']
             u_id = data['discord_id']
             uname = data['discord_username']
-            print(uname)
+            print(f"Current league: {league}, Future league: {future_league}")
             if data['league'] == data['future_league']:
                 continue
+            if data['future_league'] == None or data['league'] == None:
+                continue
+            print(f"HERE MFER          Current league: {league}, Future league: {future_league}")
+            print(f"Current league: {league}, Future league: {future_league}")
             osu_uname = await get_osu_uname(discord_uname=uname)
             pp = await get_pp(discord_username=uname)
             print(f"{osu_uname}, {pp}")
             await supabase.table(league).delete().eq('discord_username', uname).execute()
-            await supabase.table(future_league).insert({'discord_username': uname, 'osu_username': osu_uname, 'initial_pp': pp}).execute()
+            await supabase.table(future_league).insert([{
+                'discord_username': uname,
+                'osu_username': osu_uname,
+                'initial_pp': pp
+            }]).execute()
             await supabase.table('discord_osu').update({'league': future_league}).eq('discord_username', uname).execute()
             players.append({
                 'discord_username': uname,
